Move htmlSerializeEscape into its own library
